+++ R/ga_v4_get.R
@@ -17,7 +17,7 @@
 #' @param pivots Pivots of the data as created by \link{pivot_ga4}
 #' @param cohorts Cohorts created by \link{make_cohort_group}
 #' @param pageToken Where to start the data fetch
-#' @param pageSize How many rows to fetch. Max 10000 each batch.
+#' @param pageSize How many rows to fetch. Max 100000 each batch.
 #' @param samplingLevel Sample level
 #' @param metricFormat If supplying calculated metrics, specify the metric type
 #' @param histogramBuckets For numeric dimensions such as hour, a list of buckets of data.
@@ -99,6 +99,7 @@
 #' 
 #' 
 #' @family GAv4 fetch functions
+#' @import assertthat
 #' @export
 make_ga_4_req <- function(viewId,
                           date_range=NULL,
@@ -120,20 +121,21 @@ make_ga_4_req <- function(viewId,
   samplingLevel <- match.arg(samplingLevel)
   
   if(all(is.null(date_range), is.null(cohorts))){
-    stop("Must supply one of date_range or cohorts")
+    stop("Must supply one of date_range or cohorts", call. = FALSE)
   }
   
   if(!is.null(cohorts)){
-    assertthat::assert_that(cohort_metric_check(metrics),
+    assert_that(cohort_metric_check(metrics),
                             cohort_dimension_check(dimensions))
     if(!is.null(date_range)){
-      warning("Don't supply date_range when using cohorts, setting date_range to NULL")
+      warning("Don't supply date_range when using cohorts, setting date_range to NULL", 
+              call. = FALSE)
       date_range <- NULL
     }
   }
   
   if(is.null(metrics)){
-    stop("Must supply a metric")
+    stop("Must supply a metric", call. = FALSE)
   }
   
   if(!is.null(segments)){
@@ -195,33 +197,52 @@ make_ga_4_req <- function(viewId,
 
 
 
-#' Get Google Analytics v4 data (single request)
+#' Get Google Analytics v4 data
 #'
-#' A convenience function that wraps \link{make_ga_4_req} and \link{fetch_google_analytics_4}
-#'  for the common case of one GA data request.
+#' Fetch Google Analytics data using the v4 API.  For the v3 API use \link{google_analytics_3}.
 #'  
-#' Will perform automatic batching if over the 10000 row per API call limit.
+#' Will perform automatic batching if over the 100000 row per API call limit.
+#' 
+#' @section Row requests:
+#' 
+#' By default the API call will use v4 batching that splits requests into 5 seperate calls of 10k rows each.  This can go up to 100k, so this means up to 500k rows can be fetched per API call, however the API servers will fail with a 500 error if the query is too complicated as the processing time at Google's end gets too long.  In this case, you may want to tweak the \code{rows_per_call} argument downwards, or fall back to using \code{slow_fetch = FALSE} which will send an API request one at a time.  If fetching data via scheduled scripts this is recommended as the default.
+#' 
 #' 
 #' @section Anti-sampling:
 #' 
 #' \code{anti_sample} being TRUE ignores \code{max} as the API call is split over days 
 #'   to mitigate the sampling session limit, in which case a row limit won't work.  Take the top rows
 #'   of the result yourself instead e.g. \code{head(ga_data_unsampled, 50300)}
-#' 
-#' If you are lucky enough to need sub-day sampling, it will attempt to fetch per hour, but you are
-#'   restricted to not using \code{dim_filter} argument if this is the case.  
-#'   Try using \code{filtersExpression} instead.
 #'   
 #' \code{anti_sample} being TRUE will also set \code{samplingLevel='LARGE'} to minimise 
 #'   the number of calls.
+#' 
+#' @section Resource Quotas:
+#' 
+#' If you are on GA360 and have access to resource quotas,
+#'   set the \code{useResourceQuotas=TRUE} and set the Google Cloud 
+#'   client ID to the project that has resource quotas activated, 
+#'   via \link[googleAuthR]{gar_set_client} or options.
+#'   
+#' @section Caching:
+#' 
+#' By default local caching is turned on for v4 API requests.  This means that
+#'   making the same request as one this session will read from memory and not
+#'   make an API call. You can also set the cache to disk via 
+#'   the \link[googleAuthR]{gar_cache_setup} function.  This can be useful
+#'   when running RMarkdown reports using data. To empty the cache use 
+#'   \link[googleAuthR]{gar_cache_empty}.
 #' 
 #' @inheritParams make_ga_4_req
 #' @param max Maximum number of rows to fetch. Defaults at 1000. Use -1 to fetch all results. Ignored when \code{anti_sample=TRUE}.
 #' @param anti_sample If TRUE will split up the call to avoid sampling.
 #' @param anti_sample_batches "auto" default, or set to number of days per batch. 1 = daily.
-#' @param slow_fetch For large results this slows down the API requests to avoid 500 errors
-#' 
-#' @return A Google Analytics data.frame
+#' @param slow_fetch For large, complicated API requests this bypasses some API hacks that may result in 500 errors.  For smaller queries, leave this as \code{FALSE} for quicker data fetching. 
+#' @param useResourceQuotas If using GA360, access increased sampling limits. 
+#'   Default \code{NULL}, set to \code{TRUE} or \code{FALSE} if you have access to this feature. 
+#' @param rows_per_call Set how many rows are requested by the API per call, up to a maximum of 100000.
+#'   
+#' @return A Google Analytics data.frame, with attributes showing row totals, sampling etc. 
 #' 
 #' @examples 
 #' 
@@ -243,44 +264,57 @@ make_ga_4_req <- function(viewId,
 #' ga_id <- 123456
 #' 
 #' ## simple query to test connection
-#' google_analytics_4(ga_id, 
-#'                    date_range = c("2017-01-01", "2017-03-01"), 
-#'                    metrics = "sessions", 
-#'                    dimensions = "date")
+#' google_analytics(ga_id, 
+#'                  date_range = c("2017-01-01", "2017-03-01"), 
+#'                  metrics = "sessions", 
+#'                  dimensions = "date")
 #' 
 #' }
 #' 
 #' @family GAv4 fetch functions
+#' @import assertthat
 #' @export
-google_analytics_4 <- function(viewId,
-                               date_range=NULL,
-                               metrics=NULL,
-                               dimensions=NULL,
-                               dim_filters=NULL,
-                               met_filters=NULL,
-                               filtersExpression=NULL,
-                               order=NULL,
-                               segments=NULL,
-                               pivots=NULL,
-                               cohorts=NULL,
-                               max=1000,
-                               samplingLevel=c("DEFAULT", "SMALL","LARGE"),
-                               metricFormat=NULL,
-                               histogramBuckets=NULL,
-                               anti_sample = FALSE,
-                               anti_sample_batches = "auto",
-                               slow_fetch = FALSE){
+google_analytics <- function(viewId,
+                             date_range=NULL,
+                             metrics=NULL,
+                             dimensions=NULL,
+                             dim_filters=NULL,
+                             met_filters=NULL,
+                             filtersExpression=NULL,
+                             order=NULL,
+                             segments=NULL,
+                             pivots=NULL,
+                             cohorts=NULL,
+                             max=1000,
+                             samplingLevel=c("DEFAULT", "SMALL","LARGE"),
+                             metricFormat=NULL,
+                             histogramBuckets=NULL,
+                             anti_sample = FALSE,
+                             anti_sample_batches = "auto",
+                             slow_fetch = FALSE,
+                             useResourceQuotas= NULL,
+                             rows_per_call = 10000L){
   
-  # cache_call(cache)
+  timer_start <- Sys.time()
+  
+  if(!is.null(useResourceQuotas)){
+    assert_that(is.flag(useResourceQuotas))
+  }
+  
+  assert_that(is.count(rows_per_call),
+              rows_per_call <= 100000L)
+  
+  # support increase from 10k to 100k
+  api_max_rows <- rows_per_call
   
   max         <- as.integer(max)
   allResults  <- FALSE
   if(max < 0){
     ## size of 1 v4 batch 0 indexed
-    max <- as.integer(49999)
+    max <- api_max_rows - 1L
     allResults <- TRUE
   }
-  reqRowLimit <- as.integer(10000)
+  reqRowLimit <- as.integer(api_max_rows)
   
   if(!is.null(cohorts)){
     anti_sample <- FALSE
@@ -302,7 +336,8 @@ google_analytics_4 <- function(viewId,
                        metricFormat      = metricFormat,
                        histogramBuckets  = histogramBuckets,
                        anti_sample_batches = anti_sample_batches,
-                       slow_fetch = slow_fetch))
+                       slow_fetch          = slow_fetch,
+                       rows_per_call     = rows_per_call))
   }
   
   if(max > reqRowLimit){
@@ -317,7 +352,7 @@ google_analytics_4 <- function(viewId,
     start_index <- as.integer(start_index)
     
     if(allResults){
-      remaining <- as.integer(10000)
+      remaining <- as.integer(api_max_rows)
     } else {
       remaining   <- min(as.integer(max - start_index), reqRowLimit)
     }
@@ -343,11 +378,14 @@ google_analytics_4 <- function(viewId,
   
   ## if non-batching, fetch one at a time
   if(slow_fetch){
-    out <- fetch_google_analytics_4_slow(requests, max_rows = max, allRows = allResults)
+    out <- fetch_google_analytics_4_slow(requests, 
+                                         max_rows = max, allRows = allResults, 
+                                         useResourceQuotas = useResourceQuotas)
     allResults <- FALSE
   } else {
     ## only gets up to 50000 first time as we don't know true total row count yet
-    out <- fetch_google_analytics_4(requests, merge = TRUE)
+    out <- fetch_google_analytics_4(requests, 
+                                    merge = TRUE, useResourceQuotas = useResourceQuotas)
   }
 
   ## if batching, get the rest of the results now we now precise rowCount
@@ -355,12 +393,18 @@ google_analytics_4 <- function(viewId,
     all_rows <- as.integer(attr(out, "rowCount"))
     if(!is.null(out) && nrow(out) < all_rows){
       ## create the remaining requests
-      meta_batch_start_index2 <- seq(from=50000, to=all_rows, by=reqRowLimit)
+      meta_batch_start_index2 <- tryCatch(seq(from=api_max_rows, to=all_rows, by=reqRowLimit),
+                                          error = function(ex){
+                                            stop("Sequence batch error: ", 
+                                                 "\n api_max_rows: ",api_max_rows,
+                                                 "\n all_rows:", all_rows,
+                                                 "\n reqRowLimit:", reqRowLimit)
+                                          })
       ## make a list of the requests
       requests2 <- lapply(meta_batch_start_index2, function(start_index){
         
         start_index <- as.integer(start_index)
-        remaining <- as.integer(10000)
+        remaining <- as.integer(api_max_rows)
         
         make_ga_4_req(viewId            = viewId,
                       date_range        = date_range,
@@ -380,9 +424,30 @@ google_analytics_4 <- function(viewId,
                       histogramBuckets  = histogramBuckets)
         
       })
-      the_rest <- fetch_google_analytics_4(requests2, merge = TRUE)
-      out <- rbind(out, the_rest)
-      myMessage("All data downloaded, total of [",all_rows,"]", level = 3)
+      slow <- FALSE
+      the_rest <- tryCatch(fetch_google_analytics_4(requests2, 
+                                                    merge = TRUE, 
+                                                    useResourceQuotas = useResourceQuotas),
+                           error = function(ex){
+                             myMessage("Error requesting v4 batch - retrying with slow_fetch=TRUE ...", 
+                                       level = 3)
+                             slow <<- TRUE
+                             Sys.sleep(5)
+                             fetch_google_analytics_4_slow(requests, 
+                                                           max_rows = max, 
+                                                           allRows = allResults, 
+                                                           useResourceQuotas=useResourceQuotas)
+                           })
+      
+      if(!slow){
+        out <- rbind(out, the_rest)
+      } else {
+        out <- the_rest
+      }
+
+      myMessage("All data downloaded, total of [",all_rows,"] - download time: ",
+                format(Sys.time() - timer_start), 
+                level = 3)
       
     } else {
       myMessage("One batch enough to get all results", level = 1)
@@ -398,6 +463,14 @@ google_analytics_4 <- function(viewId,
 
 }
 
+#' @rdname google_analytics
+#' @param ... Arguments passed to \link{google_analytics}
+#' @export
+google_analytics_4 <- function(...){
+  .Deprecated("google_analytics", package = "googleAnalyticsR")
+  google_analytics(...)
+}
+
 #' Fetch GAv4 requests one at a time
 #' 
 #' Due to large complicated queries causing the v4 API to timeout, 
@@ -406,11 +479,16 @@ google_analytics_4 <- function(viewId,
 #' @param request_list A list of requests created by \link{make_ga_4_req}
 #' @param max_rows Number of rows requested (if not fetched)
 #' @param allRows Whether to fetch all available rows
+#' @param useResourceQuotas If using GA360, access increased sampling limits. 
+#'   Default \code{NULL}, set to \code{TRUE} or \code{FALSE} if you have access to this feature. 
 #' 
 #' @return A dataframe of all the requests
 #' @importFrom googleAuthR gar_api_generator
 #' @family GAv4 fetch functions
-fetch_google_analytics_4_slow <- function(request_list, max_rows, allRows = FALSE){
+fetch_google_analytics_4_slow <- function(request_list, 
+                                          max_rows, 
+                                          allRows = FALSE, 
+                                          useResourceQuotas=NULL){
   
   ## make the fetch function
   myMessage("Calling APIv4 slowly....", level = 2)
@@ -434,8 +512,11 @@ fetch_google_analytics_4_slow <- function(request_list, max_rows, allRows = FALS
   while(do_it){
     
     body <- list(
-      reportRequests = the_req
+      reportRequests = the_req,
+      useResourceQuotas = useResourceQuotas
     )
+    
+    body <- rmNullObs(body)
     
     myMessage("Slow fetch: [", 
               the_req$pageToken, "] from estimated actual Rows [", actualRows, "]", 
@@ -469,7 +550,9 @@ fetch_google_analytics_4_slow <- function(request_list, max_rows, allRows = FALS
 #'
 #' @param request_list A list of requests created by \link{make_ga_4_req}
 #' @param merge If TRUE then will rbind that list of data.frames
-#'
+#' @param useResourceQuotas If using GA360, access increased sampling limits. 
+#'   Default \code{NULL}, set to \code{TRUE} or \code{FALSE} if you have access to this feature. 
+#'     
 #' @return A dataframe if one request, or a list of data.frames if multiple.
 #'
 #' @importFrom googleAuthR gar_api_generator
@@ -505,10 +588,11 @@ fetch_google_analytics_4_slow <- function(request_list, max_rows, allRows = FALS
 #' }
 #' 
 #' @family GAv4 fetch functions
+#' @import assertthat
 #' @export
-fetch_google_analytics_4 <- function(request_list, merge = FALSE){
+fetch_google_analytics_4 <- function(request_list, merge = FALSE, useResourceQuotas = NULL){
 
-  assertthat::assert_that(is.list(request_list))
+  assert_that(is.list(request_list))
   ## amount of batches per v4 api call
   ga_batch_limit <- 5
   
@@ -537,7 +621,6 @@ fetch_google_analytics_4 <- function(request_list, merge = FALSE){
   f <- gar_api_generator("https://analyticsreporting.googleapis.com/v4/reports:batchGet",
                          "POST",
                          data_parse_function = google_analytics_4_parse_batch,
-                         # data_parse_function = function(x) x,
                          simplifyVector = FALSE)
   
   ## if under 5, one call
@@ -546,8 +629,11 @@ fetch_google_analytics_4 <- function(request_list, merge = FALSE){
     request_list <- unitToList(request_list)
     
     body <- list(
-      reportRequests = request_list
+      reportRequests = request_list,
+      useResourceQuotas = useResourceQuotas
     )
+    
+    body <- rmNullObs(body)
 
     out <- try(f(the_body = body))
     
@@ -562,14 +648,20 @@ fetch_google_analytics_4 <- function(request_list, merge = FALSE){
                          function(x) request_list[x:(x+(ga_batch_limit-1))])
     
     ## make the body for each v4 api call
-    body_list <- lapply(batch_list, function(x) list(reportRequests = x))
+    body_list <- lapply(batch_list, function(x){
+      bb <- list(reportRequests = x, 
+                 useResourceQuotas = useResourceQuotas)
+      rmNullObs(bb)
+    })
+    
     body_list <- rmNullObs(body_list)
       
     ## loop over the requests normally
-    myMessage("Looping over maximum [", length(body_list), "] batches.", level = 1)
+    myMessage("Looping over maximum [", length(body_list), "] batches.", level = 1) 
+    
     response_list <- lapply(body_list, function(b){
-      
-      myMessage("Fetching data batch...", level = 3)
+
+      myMessage("Fetching v4 data batch...", level = 3)
       
       out <- try(f(the_body = b))