+++ R/options.R
@@ -11,7 +11,8 @@
                                     "https://www.googleapis.com/auth/analytics.readonly",
                                     "https://www.googleapis.com/auth/analytics.manage.users.readonly",
                                     "https://www.googleapis.com/auth/analytics.edit",
-                                    "https://www.googleapis.com/auth/analytics.manage.users")
+                                    "https://www.googleapis.com/auth/analytics.manage.users"),
+    googleAuthR.httr_oauth_cache = "ga.oauth"
   )
   
   toset <- !(names(op.googleAnalyticsR) %in% names(op))
@@ -20,7 +21,8 @@
   
   ## override existing options
   options(
-    googleAuthR.batch_endpoint = "https://www.googleapis.com/batch/analytics/v3")
+    googleAuthR.batch_endpoint = "https://www.googleapis.com/batch/analytics/v3",
+    googleAuthR.tryAttempts = 1)
   
   if(Sys.getenv("GA_CLIENT_ID") != ""){
     options(googleAuthR.client_id = Sys.getenv("GA_CLIENT_ID"))
@@ -41,7 +43,10 @@
   default_project_message()
   
   f <- function(req){
-    if(!is.null(req$content$reports)){
+    
+    stuff <- tryCatch(req$content$reports, error = function(x) NULL)
+    
+    if(!is.null(stuff)){
       return(TRUE)
     } else {
       return(FALSE)