+++ data-raw/get_meta_data.R
@@ -1,6 +1,6 @@
 ## script to get meta data, rerun each release
 library(googleAnalyticsR)
-ga_auth()
+#ga_auth()
 meta <- google_analytics_meta()
 devtools::use_data(meta, overwrite = TRUE)
 devtools::use_data_raw()